requirements.txt
# Core dependencies
anthropic>=0.18.0
openai>=1.12.0
google-generativeai>=0.3.0
requests>=2.31.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community

# LLM Providers
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-ollama>=0.1.0

# Embeddings
langchain-huggingface
sentence-transformers>=2.2.2

# Vector Database
langchain-chroma
chromadb

# Utilities
python-dotenv==1.0.0

# Metrics calculation
nltk>=3.8.0
sacrebleu>=2.3.0
bert-score>=0.3.13

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# API Bridge
fastapi
uvicorn[standard]
pydantic

# Optional: For advanced metrics
# sentence-transformers>=2.2.0
# transformers>=4.30.0